+++ src/modules/artist/artist.service.ts
@@ -7,20 +7,20 @@ import { PrismaService } from '../../../prisma/prisma.service';
 
 @Injectable()
 export class ArtistService {
-  constructor (private readonly prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) {}
   async getAllArtist(): Promise<IArtist[]> {
     return this.prisma.artist.findMany();
   }
 
   createArtist(createArtistDto: CreateArtistDto): Promise<IArtist> {
-    return this.prisma.artist.create({ data: createArtistDto});
+    return this.prisma.artist.create({ data: createArtistDto });
   }
 
   async getArtistsById(id: string): Promise<IArtist> {
     const artist = await this.prisma.artist.findUnique({
       where: {
-        id
-      }
+        id,
+      },
     });
 
     if (!artist) {
@@ -30,11 +30,14 @@ export class ArtistService {
     return artist;
   }
 
-  async updateArtistById(updateArtistDto: UpdateArtistDto, id: string): Promise<IArtist> {
+  async updateArtistById(
+    updateArtistDto: UpdateArtistDto,
+    id: string,
+  ): Promise<IArtist> {
     const artist = await this.prisma.artist.findUnique({
       where: {
-        id
-      }
+        id,
+      },
     });
 
     if (!artist) {
@@ -43,21 +46,45 @@ export class ArtistService {
 
     return this.prisma.artist.update({
       where: {
-        id
+        id,
       },
       data: {
         ...updateArtistDto,
-        id
-      }
+        id,
+      },
     });
   }
 
   async deleteArtistById(id: string) {
+    const artist = await this.prisma.artist.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!artist) {
+      throw new NotFoundException('Artist not found');
+    }
+
     await this.prisma.artist.delete({
       where: {
-        id
-      }
+        id,
+      },
     });
+
+    const favorites = await this.prisma.favorites.findFirst();
+
+    await this.prisma.favorites.update({
+      where: {
+        id: favorites.id,
+      },
+      data: {
+        artists: {
+          set: favorites.artists.filter((ids) => ids !== id),
+        },
+      },
+    });
+
     return;
   }
 }
